+++ script/SafeConfig.sol
@@ -24,6 +24,7 @@ library SafeConfig {
                 || block.chainid == 480 // worldchain
                 || block.chainid == 999 // hyperevm
                 || block.chainid == 2741 // abstract
+                || block.chainid == 4217 // tempo
                 || block.chainid == 5000 // mantle
                 || block.chainid == 8453 // base
                 || block.chainid == 9745 // plasma
@@ -58,6 +59,7 @@ library SafeConfig {
                 || block.chainid == 480 // worldchain
                 || block.chainid == 999 // hyperevm
                 || block.chainid == 2741 // abstract
+                || block.chainid == 4217 // tempo
                 || block.chainid == 5000 // mantle
                 || block.chainid == 8453 // base
                 || block.chainid == 9745 // plasma
@@ -97,6 +99,7 @@ library SafeConfig {
                 || block.chainid == 146 // sonic
                 || block.chainid == 480 // worldchain
                 || block.chainid == 999 // hyperevm
+                || block.chainid == 4217 // tempo
                 || block.chainid == 5000 // mantle
                 || block.chainid == 8453 // base
                 || block.chainid == 9745 // plasma