+++ src/chains/Tempo/BridgeSettler.sol
@@ -1,14 +1,14 @@
 // SPDX-License-Identifier: MIT
 pragma solidity =0.8.33;
 
-import {IERC20} from "@forge-std/interfaces/IERC20.sol";
 import {SettlerAbstract} from "../../SettlerAbstract.sol";
 import {IBridgeSettlerActions} from "../../bridge/IBridgeSettlerActions.sol";
 import {BridgeSettler, BridgeSettlerBase} from "../../bridge/BridgeSettler.sol";
+import {Across} from "../../core/Across.sol";
 
-contract KatanaBridgeSettler is BridgeSettler {
+contract TempoBridgeSettler is BridgeSettler, Across {
     constructor(bytes20 gitCommit) BridgeSettlerBase(gitCommit) {
-        assert(block.chainid == 747474 || block.chainid == 31337);
+        assert(block.chainid == 4217 || block.chainid == 31337);
     }
 
     function _dispatch(uint256 i, uint256 action, bytes calldata data)
@@ -18,6 +18,12 @@ contract KatanaBridgeSettler is BridgeSettler {
     {
         if (super._dispatch(i, action, data)) {
             return true;
+        } else if (action == uint32(IBridgeSettlerActions.BRIDGE_ERC20_TO_ACROSS.selector)) {
+            (address spoke, bytes memory depositData) = abi.decode(data, (address, bytes));
+            bridgeERC20ToAcross(spoke, depositData);
+        } else if (action == uint32(IBridgeSettlerActions.BRIDGE_NATIVE_TO_ACROSS.selector)) {
+            (address spoke, bytes memory depositData) = abi.decode(data, (address, bytes));
+            bridgeNativeToAcross(spoke, depositData);
         } else {
             return false;
         }