+++ src/chains/Tempo/TakerSubmitted.sol
@@ -1,14 +1,12 @@
 // SPDX-License-Identifier: MIT
 pragma solidity =0.8.33;
 
-import {KatanaMixin} from "./Common.sol";
+import {TempoMixin} from "./Common.sol";
 import {Settler} from "../../Settler.sol";
 
 import {IERC20} from "@forge-std/interfaces/IERC20.sol";
 import {ISignatureTransfer} from "@permit2/interfaces/ISignatureTransfer.sol";
 import {ISettlerActions} from "../../ISettlerActions.sol";
-import {FastLogic} from "../../utils/FastLogic.sol";
-import {revertConfusedDeputy} from "../../core/SettlerErrors.sol";
 
 // Solidity inheritance is stupid
 import {SettlerAbstract} from "../../SettlerAbstract.sol";
@@ -17,7 +15,7 @@ import {Permit2PaymentAbstract} from "../../core/Permit2PaymentAbstract.sol";
 import {AbstractContext} from "../../Context.sol";
 
 /// @custom:security-contact security@0x.org
-contract KatanaSettler is Settler, KatanaMixin {
+contract TempoSettler is Settler, TempoMixin {
     constructor(bytes20 gitCommit) SettlerBase(gitCommit) {}
 
     function _dispatchVIP(uint256 action, bytes calldata data) internal override DANGEROUS_freeMemory returns (bool) {
@@ -47,7 +45,7 @@ contract KatanaSettler is Settler, KatanaMixin {
 
     function _dispatch(uint256 i, uint256 action, bytes calldata data)
         internal
-        override(Settler, KatanaMixin)
+        override(Settler, TempoMixin)
         returns (bool)
     {
         return super._dispatch(i, action, data);
@@ -55,5 +53,14 @@ contract KatanaSettler is Settler, KatanaMixin {
 
     function _msgSender() internal view override(Settler, AbstractContext) returns (address) {
         return super._msgSender();
+    }
+
+    function _fallback(bytes calldata data)
+        internal
+        virtual
+        override(Permit2PaymentAbstract, TempoMixin)
+        returns (bool, bytes memory)
+    {
+        return super._fallback(data);
     }
 }